app/requirements.txt
kafka-python-ng==2.2.3
elasticsearch
pydantic